pyproject.toml
[project]
name = "traffic-oracle"
version = "0.1.0"
description = "OpenEnv environment for traffic signal optimization via RL"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "setuptools",
    "wheel",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["server*", "baseline*", "tests*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
